+++ Loenn/entities/vol2/dream_field.lua
@@ -2,7 +2,7 @@ local dreamBlock = {}
 
 dreamBlock.name = "ScugHelper/DreamField"
 dreamBlock.fillColor = {0.0, 0.0, 0.0, 0.2}
-dreamBlock.borderColor = {0, 0, 0, 0}
+dreamBlock.borderColor = {1, 1, 1, 0.1}
 dreamBlock.nodeLineRenderType = "line"
 dreamBlock.nodeLimits = { 0, 1 }
 dreamBlock.depth = 8400